+++ CHANGELOG.md
@@ -1,9 +1,22 @@
+# Version 0.25.0 - 2022-05-31
+
+* github-184: Only read in *.yml / *.yaml files in module loader
+* github-183: Support storing SQL in external file in `hiveView`
+* github-185: Missing _SUCCESS file when writing to dynamic partitions
+* github-186: Support output mode `OVERWRITE_DYNAMIC` for Delta relation
+* github-149: Support creating views in JDBC with new `jdbcView` relation
+* github-190: Replace logo in documentation
+* github-188: Log detailed timing information when writing to JDBC relation
+* github-191: Add user provided description to quality checks
+* github-192: Provide example queries for JDBC metric sink
+
+
 # Version 0.24.1 - 2022-04-28
 
 * github-175: '--jobs' parameter starts way to many parallel jobs
 * github-176: start-/end-date in report should not be the same
 * github-177: Implement generic SQL schema check
-* github-179: Update DeltaLake dependency to 1.2.0
+* github-179: Update DeltaLake dependency to 1.2.1
 
 
 # Version 0.24.0 - 2022-04-05